Add files via upload
This activity was definitely challenging once it got to step four. I was trying to do too much and make a for function to try and go through the list of names, however a simpler approach is what was required. In the end I learned how to call objects within an array to be used within the props element.