Add files via upload
Worked on arrow functions. Restructuring older pieces of code into arrow functions. 6 challenges present. I'm getting the hang of it. I may still write it the original way at first. But my goal is to go back and begin converting older functions in past projects into arrow functions if they can be converted as such. Truly though an arrow function should be used on anonymous functions when the function does not need to be defined so an arrow function may not be appropriate for all use cases.