Work and Solutions for Event Handling
useState property is used to create a state of true or false to determine whether or not the button was being moused over. In the button itself, we created an inline style that use a ternary method to check for the state of isMousedOver. True to black, false to white. To ensure that we set the state of true and false multiple times we need to introduce the html onMouseOut event listener. Inside of its function it also references the setMouseOver hook from earlier to reset the state to false when the button is not being hovered. 

While my original solution does work, it does not create a state which can be read and interpreted later. It is reactive without the ability to be interpreted later on down the road.