Updating Form with Objects
Learned how to create an object while using states. Form now updates the "prevValue" to the newValue which is equal to event.target.value. The value unaffected is taken not of. Example. If the input occurs on the first name, last name will be set to prevValue.lName. It references that portion of the object whether it is blank or it has been modified with a new newValue. inputName logs which field is being used as its event targets the "name" of the input. fName and lName. That is why we are setting the conditional statement to read the input name first in order to decide what needs to be changed and what needs to be left alone.